# dienhoatructuyen2424-be/cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "API DOCUMENTATION PARKING BACKEND",
        "termsOfService": "github.com/phananhtu1998/go-backend-api",
        "contact": {
            "name": "TEAM AI",
            "url": "github.com/phananhtu1998/go-backend-api",
            "email": "teamaithacoindustries@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/v1/2025",
    "paths": {
...
        "/funcpackage/create_func_package": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Api tạo gói chức năng trong hệ thống",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FuncPackages"
                ],
                "summary": "Tạo gói chức năng",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FuncpackageInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/funcpackage/delete_funcpackage/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API này xóa gói chức năng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FuncPackages"
                ],
                "summary": "Xóa gói chức năng",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID của gói chức năng cần xóa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/funcpackage/get_all_func_package": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Api lấy gói chức năng trong hệ thống",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FuncPackages"
                ],
                "summary": "Lấy gói chức năng",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Từ khóa tìm kiếm",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Số trang (mặc định: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Số lượng mỗi trang (mặc định: 20)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/funcpackage/get_all_func_package_select": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Api lấy gói chức năng cho select trong hệ thống",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FuncPackages"
                ],
                "summary": "Lấy gói chức năng cho select",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
...
        "/license/create_license": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API tạo mới license cho gói chức năng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "License"
                ],
                "summary": "Tạo mới license cho gói chức năng",
                "parameters": [
                    {
                        "description": "Thông tin License cần tạo",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.License"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/license/license_renewal/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API gia hạn license",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "License"
                ],
                "summary": "Gia hạn license",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID license cần gia hạn",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dữ liệu gia hạn license",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LicenseRenewal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.FuncpackageInput": {
            "type": "object",
            "properties": {
                "role_max_number": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "model.License": {
            "type": "object",
            "required": [
                "dateend",
                "role_id"
            ],
            "properties": {
                "dateend": {
                    "type": "string",
                    "example": "2025-03-30 11:22:30"
                },
                "datestart": {
                    "type": "string",
                    "example": "2025-03-28 11:22:30"
                },
                "role_id": {
                    "type": "string",
                    "example": "2b796313-1134-44b3-b527-2c27d41a1624"
                }
            }
        },
        "model.LicenseRenewal": {
            "type": "object",
            "required": [
                "dateend"
            ],
            "properties": {
                "dateend": {
                    "type": "string",
                    "example": "2025-03-30 11:22:30"
                },
                "datestart": {
                    "type": "string",
                    "example": "2025-03-28 11:22:30"
                }
            }
        },
...
        "response.ErrorResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "status code",
                    "type": "integer"
                },
                "detail": {
                    "description": "dữ liệu được return"
                },
                "error": {
                    "description": "Thông báo lỗi",
                    "type": "string"
                }
            }
        },
        "response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "status code",
                    "type": "integer"
                },
                "data": {
                    "description": "dữ liệu được return"
                },
                "message": {
                    "description": "Thông báo lỗi",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Nhập API Key vào header để xác thực API",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Nhập Bearer Token vào header. Ví dụ: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TokenKey": {
            "description": "Nhập Token Key vào header để xác thực API",
            "type": "apiKey",
            "name": "TokenKey",
            "in": "header"
        }
    },
    "security": [
        {
            "": [
                "BearerAuth",
                "ApiKeyAuth",
                "TokenKey"
            ]
        }
    ]
}