# dienhoatructuyen2424-be/internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package database

import (
	"database/sql"
	"encoding/json"
	"time"
)

// account
type Account struct {
	ID       string
	Number   int32
	Code     string
	Name     string
	Email    string
	Username string
	Password string
	Salt     string
	// [active,inactive]
	Status    bool
	Images    string
	CreatedBy string
	IsDeleted bool
	CreateAt  time.Time
	UpdateAt  time.Time
}

type Area struct {
	ID        string
	Code      string
	Name      string
	IsDeleted bool
	CreatedAt time.Time
	UpdatedAt time.Time
}

type CustomerBase struct {
	ID        string
	Number    int32
	Email     string
	Password  string
	Salt      string
	CreatedAt time.Time
	UpdateAt  time.Time
}

type CustomerInfo struct {
	ID        string
	Code      string
	Username  string
	Name      string
	PaymentID string
	Image     string
	// state: 0-Locked, 1-Activated, 2-Not Activated
	State int8
	// Authentication status: 0-Not Authenticated, 1-Pending, 2-Authenticated, 3-Failed
	IsAuthentication int8
	Mobile           string
	Gender           int8
	Birthday         time.Time
	Email            string
	CreatedAt        time.Time
	UpdateAt         time.Time
}

type CustomerVerify struct {
	ID            string
	VerifyOtp     string
	VerifyEmail   string
	VerifyKeyHash string
	IsVerified    bool
	IsDeleted     bool
	CreatedAt     time.Time
	UpdateAt      time.Time
}

// keytoken
type Keytoken struct {
	ID                string
	AccountID         string
	RefreshToken      string
	RefreshTokensUsed json.RawMessage
	CreateAt          time.Time
	UpdateAt          time.Time
}

// keytokencustomer
type Keytokencustomer struct {
	ID                string
	CustomerID        string
	RefreshToken      string
	RefreshTokensUsed json.RawMessage
	CreateAt          time.Time
	UpdateAt          time.Time
}

type License struct {
	ID        string
	License   string
	RoleID    string
	DateStart time.Time
	DateEnd   string
	CreatedAt time.Time
	UpdateAt  time.Time
	IsDeleted bool
}

type ManageBookingList struct {
	ID          string
	Code        string
	CustomerID  string
	TimeBooking time.Time
	Description string
	Status      string
	Note        string
	IsDeleted   bool
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type Menu struct {
	ID              string
	MenuName        string
	MenuIcon        string
	MenuUrl         string
	MenuParentID    sql.NullString
	MenuLevel       int32
	MenuNumberOrder int32
	MenuGroupName   string
	IsDeleted       bool
	CreateAt        time.Time
	UpdateAt        time.Time
}

type ParkingFloorManagement struct {
	ID           string
	Name         string
	ParkingLotID string
	Number       int32
	Column       int32
	Row          int32
	IsDeleted    bool
	CreatedAt    time.Time
	UpdatedAt    time.Time
}

type ParkingLot struct {
	ID                    string
	Code                  string
	Name                  string
	AreaID                string
	NumberOfFloors        int32
	TotalParkingPositions int32
	Status                int8
	IsDeleted             bool
	CreatedAt             time.Time
	UpdatedAt             time.Time
}

type ParkingManagement struct {
	ID            string
	Code          string
	CustomerID    string
	ParkingSpotID string
	Status        int8
	TimeCheckin   sql.NullString
	CreatedAt     time.Time
	UpdatedAt     time.Time
}

type ParkingSpot struct {
	ID             string
	Name           string
	ParkingFloorID string
	Qrcode         string
	SpotNumber     int32
	X              int32
	Y              int32
	Status         int8
	IsDeleted      bool
	CreatedAt      time.Time
	UpdatedAt      time.Time
}

type Payment struct {
	ID        string
	Name      string
	IsDeleted bool
	CreatedAt time.Time
	UpdatedAt time.Time
}

type Role struct {
	ID             string
	Code           string
	RoleName       string
	RoleLeftValue  int32
	RoleRightValue int32
	RoleMaxNumber  int32
	LicenseID      string
	CreatedBy      string
	IsDeleted      bool
	CreateAt       time.Time
	UpdateAt       time.Time
}

type RoleAccount struct {
	ID        string
	AccountID string
	RoleID    string
	CreatedBy sql.NullString
	IsDeleted bool
	CreateAt  time.Time
	UpdateAt  time.Time
}

type RolesMenu struct {
	ID         string
	MenuID     string
	RoleID     string
	ListMethod json.RawMessage
	IsDeleted  bool
	CreateAt   time.Time
	UpdateAt   time.Time
}

type TransactionHistory struct {
	ID          string
	NumberPlate string
	TimeIn      time.Time
	TimeOut     time.Time
	SpotNumber  int32
	Status      bool
	IsDeleted   bool
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type Vehicle struct {
	ID           string
	Code         string
	NumberPlate  string
	CustomerID   string
	ParkingLotID string
	IsDeleted    bool
	CreatedAt    time.Time
	UpdatedAt    time.Time
}